+++ packages/af-utils-react-mobx-table/src/constants/index.js
@@ -1,2 +1,1 @@
 export const HEADER_DND_TYPE = "h";
-export const EMPTY_ARRAY = [];