+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtual/index.js
@@ -1,4 +1,5 @@
 import { useEffect } from "react";
+import getEstimatedItemSizeDefault from "utils/getEstimatedItemSize";
 import useVirtualModel from "../useVirtualModel";
 
 const useVirtual = params => {
@@ -11,7 +12,10 @@ const useVirtual = params => {
         */
         model._startBatch();
         model.setHorizontal(params.horizontal);
-        model.setItemCount(params.itemCount);
+        model.setItemCount(
+            params.itemCount,
+            params.getEstimatedItemSize || getEstimatedItemSizeDefault
+        );
         model._endBatch();
     });