+++ packages/af-utils-react-virtual-headless/src/utils/getEstimatedItemSize/index.js
@@ -1,0 +1,4 @@
+const getEstimatedItemSize = (oldItemSizes, oldScrollSize) =>
+    oldItemSizes.length ? Math.round(oldScrollSize / oldItemSizes.length) : 40;
+
+export default getEstimatedItemSize;