+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @af-utils/react-mobx-table
 
+## 0.0.8
+
+### Patch Changes
+
+-   changed estimatedItemSize logics + upgraded deps
+-   Updated dependencies
+    -   @af-utils/react-table@0.0.8
+    -   @af-utils/react-virtual-headless@0.0.7
+    -   @af-utils/styled@0.0.4
+
 ## 0.0.7
 
 ### Patch Changes