+++ packages/af-utils-styled/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/styled
 
+## 0.0.4
+
+### Patch Changes
+
+-   changed estimatedItemSize logics + upgraded deps
+
 ## 0.0.3
 
 ### Patch Changes